requirements.txt
polib==1.2.0
openai>=1.3.0
python-dotenv==1.0.0 